Battery voltage detection and Blink without delay
The Battery voltage detection part of the code has been changed as instructed. The battery voltage measurement is no longer done in the loop but inside the Low Battery Voltage function. A new LED blink function is introduced instead of the existing LED blinking which uses delay(). Now millis() method is used instead of delay().